Update Magic8Ball.py
Added GUI. Having trouble with the rendering of buttons, installed python 3.7 with homebrew, seems this is a common issue with that method. However they buttons are still in position and work, going to reinstall python and run again to see how it will look. Can to clean up GUI once I can view it fully